Add getEstimateTranslation utility
