+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Utility `getEstimateTranslation` function.
+
 ## [1.1.6] - 2020-12-17
 
 ### Added
